refactor(admin): fragment god file into focused domain modules
admin.fn.ts was 764 lines with 16 server functions covering users,
bugs, challenges, submissions, tutorials, achievements, newsletter,
messages, content sync, and stats.

- Split into 10 focused modules under src/server/admin/
- Keep admin.fn.ts as thin backward-compatibility barrel
- Update all admin route imports
- Each module now has clear locality for its domain

Closes #193
