+++ src/routes/admin/messages.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { getAdminMessages, updateMessageStatus } from '@/server/admin.fn';
+import { getAdminMessages, updateMessageStatus } from '@/server/admin/messages.fn';
 import { Card, CardContent } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
 import { ArrowLeft, Mail } from 'lucide-react';
