+++ src/routes/admin/submissions.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { useQuery } from '@tanstack/react-query';
-import { getAdminSubmissions } from '@/server/admin.fn';
+import { getAdminSubmissions } from '@/server/admin/submissions.fn';
 import { Card, CardContent } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
 import {
