+++ src/routes/admin/tutorials.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { getAdminTutorials, updateTutorialStatus } from '@/server/admin.fn';
+import { getAdminTutorials, updateTutorialStatus } from '@/server/admin/tutorials.fn';
 import { Card, CardContent } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
 import { BookOpen, Search, Eye, Clock } from 'lucide-react';
